10 commit - tests + some updates

+++ src/main/java/pl/coderslab/assessmentsheetgpt/team/Team.java
@@ -1,5 +1,6 @@
 package pl.coderslab.assessmentsheetgpt.team;
 
+import com.fasterxml.jackson.datatype.jsr310.deser.key.ZonedDateTimeKeyDeserializer;
 import lombok.*;
 import pl.coderslab.assessmentsheetgpt.monitor.Monitor;
 
@@ -17,7 +18,7 @@ public class Team {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false)
     private String name;
@@ -28,9 +29,4 @@ public class Team {
     @ManyToOne
     @ToString.Exclude
     private Monitor monitor;
-
-
-
-
-
 }
